change dtype and add flexibility to batch prompts

+++ src/main/java/ai/nets/samj/ui/ConsumerInterface.java
@@ -74,7 +74,7 @@ public abstract class ConsumerInterface {
 
 	public abstract < T extends RealType< T > & NativeType< T > > RandomAccessibleInterval<T> getFocusedImageAsRai();
 	
-	public abstract List<long[]> getPointRoisOnFocusImage();
+	public abstract List<int[]> getPointRoisOnFocusImage();
 	
 	public abstract List<Rectangle> getRectRoisOnFocusImage();
